LPS-151251 Move parameter related methods from HttpComponentsUtil to URIBuilder
